Extract common views to BaseState
